extracted  function from component
